pyproject.toml
[project]
name = "IdeaAgent"
version = "0.1.0"
description = "Experimental Agent for validating machine learning research ideas"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "IdeaAgent Team" }
]
dependencies = [
    "click>=8.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "pyyaml>=6.0",
    "strictyaml>=1.7.3",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.25.0",
    "mcp>=1.0.0",
    "fastmcp>=0.1.0",
]

[project.scripts]
IdeaAgent = "ideaagent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ideaagent"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.8.0",
    "black>=23.0.0",
]
